Fix batch size capping and epochs float bug
- Cap batch_size at target_eff_batch to avoid exceeding target on
  large GPUs (e.g., A100 80GB was getting eff=256 instead of 128)
- Cast num_epochs to int (epochs_multiplier=1.5 caused float error)

Verified all models on 24GB/40GB/80GB GPUs now match target effective
batch sizes exactly.

+++ scripts/train_unified.py
@@ -143,7 +143,7 @@ model_cfg = get_model_config(model_size, vocab_size, context_size, base_epochs)
 
 # Apply epochs multiplier if present (e.g., small-long uses 2x epochs)
 epochs_multiplier = model_cfg.get("epochs_multiplier", 1)
-num_epochs = base_epochs * epochs_multiplier
+num_epochs = int(base_epochs * epochs_multiplier)
 
 # Compute init
 device_type = autodetect_device_type() if device_type == "" else device_type
